# Optimize `fill_bytes_via_*` and prepare rand_core 0.1