--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,7 +29,7 @@ serde1 = ["serde", "serde_derive", "rand_core/serde1"] # enables serialization f
 members = ["rand_core"]
 
 [dependencies]
-rand_core = { path="rand_core", default-features = false }
+rand_core = { path = "rand_core", version = "0.1.0-pre.0", default-features = false }
 log = { version = "0.4", optional = true }
 serde = { version = "1", optional = true }
 serde_derive = { version = "1", optional = true }