Benchmark before:
```
test gen_bytes_xorshift    ... bench:   1,131,259 ns/iter (+/- 2,399) = 905 MB/s
```

After:
```
test gen_bytes_xorshift    ... bench:     284,238 ns/iter (+/- 906) = 3602 MB/s
```

I updated the version number, changelog and readme of `rand_core` for a 0.1.0 release. I hope https://github.com/rust-lang-nursery/rand/pull/383 and https://github.com/rust-lang-nursery/rand/pull/396 can me merged first. The date in the changelog is set to 2018-04-15, this Sunday. @dhardy Do you think that is realistic?