@pitdicker 2018-04-13
So I had multiple attempts to improve the performance of `fill_via_u*` lying around, and forgot which one it was that did the trick. And of course I took the one with the typo that was incorrect and because of that seemingly 4x faster...

Got the real one, which is less spectacular but still ~2,5x faster (on both x86 and x86_64).
It uses both `next_u64` and `next_u32`, so I just renamed it to `fill_via_next`.

```
test gen_bytes_xorshift    ... bench:     391,250 ns/iter (+/- 27,374) = 2617 MB/s
test gen_u32_xorshift      ... bench:         860 ns/iter (+/- 36) = 4651 MB/s
test gen_u64_xorshift      ... bench:       1,722 ns/iter (+/- 273) = 4645 MB/s
```