@pitdicker 2018-04-13
> I'm not quite sure about the rand-core release; either we just make a 0.1.0-pre.1 for now or we wait a few more days (preferably with separate PR).

I am a bit inpatient for a release to be honest. But I suppose it is also my fault as I keep making changes to `rand_core`...

I am not really a fan of a second pre-release, but will leave things to you. Should I separate out the version bump from this PR?