@dhardy 2018-04-13
I seem to remember you were not a fan of the 0.5 pre-release either, yet we learned quite a bit since then...

Yes, separate the PR then we can merge the first bit and sit on the potential release a couple of days (maybe do that Monday).