@pitdicker 2018-04-13
> I seem to remember you were not a fan of the 0.5 pre-release either, yet we learned quite a bit since then...

True, I'll keep a bit more quite :smile:.

But for `rand_core` I don't remember much more than https://github.com/rust-lang-nursery/rand/issues/354 and https://github.com/rust-lang-nursery/rand/issues/378, which both seem minor documentation issues to me.