@pitdicker 2018-04-13
Ok, I reset the version to `0.1.0-pre.0`.